Extract lock file validation into plugin.yarn-lock-validation.gradle.kts
